requirements.txt
PyQt5>=5.15.0
pywin32>=300  # 新增，用于Windows注册表操作
Pillow>=9.0.0  # 新增，用于创建默认图标